requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-cors
python-dotenv
flask-jwt-extended
